+++ src/client.ts
@@ -14,6 +14,7 @@ import { WSConnectionFallback } from './connection_fallback';
 import { isErrorResponse, isWSFailure } from './errors';
 import {
   addFileToFormData,
+  axiosParamsSerializer,
   chatCodes,
   isFunction,
   isOnline,
@@ -50,6 +51,7 @@ import {
   ChannelSort,
   ChannelStateOptions,
   CheckPushResponse,
+  CheckSNSResponse,
   CheckSQSResponse,
   Configs,
   ConnectAPIResponse,
@@ -146,6 +148,7 @@ import {
   TaskStatus,
   TestCampaignResponse,
   TestPushDataInput,
+  TestSNSDataInput,
   TestSQSDataInput,
   TokenOrProvider,
   UnBanUserOptions,
@@ -155,6 +158,7 @@ import {
   UpdateCommandResponse,
   UpdatedMessage,
   UpdateMessageAPIResponse,
+  UpdateMessageOptions,
   UserCustomEvent,
   UserFilters,
   UserOptions,
@@ -303,6 +307,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
 
     this.defaultWSTimeoutWithFallback = 6000;
     this.defaultWSTimeout = 15000;
+
+    this.axiosInstance.defaults.paramsSerializer = axiosParamsSerializer;
 
     /**
      * logger function should accept 3 parameters:
@@ -712,6 +718,20 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   /**
+   * testSNSSettings - Tests that the given or configured SNS configuration is valid
+   *
+   * @param {TestSNSDataInput} [data] Overrides SNS settings for testing if needed
+   *  IE: {
+        sns_key: 'auth_key',
+        sns_secret: 'auth_secret',
+        sns_topic_arn: 'topic_to_publish_to',
+      }
+   */
+  async testSNSSettings(data: TestSNSDataInput = {}) {
+    return await this.post<CheckSNSResponse>(this.baseURL + '/check_sns', data);
+  }
+
+  /**
    * Disconnects the websocket and removes the user from client.
    *
    * @param timeout Max number of ms, to wait for close event of websocket, before forcefully assuming successful disconnection.
@@ -948,6 +968,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
         case 'post':
           response = await this.axiosInstance.post(url, data, requestConfig);
           break;
+        case 'postForm':
+          response = await this.axiosInstance.postForm(url, data, requestConfig);
+          break;
         case 'put':
           response = await this.axiosInstance.put(url, data, requestConfig);
           break;
@@ -1011,10 +1034,10 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     contentType?: string,
     user?: UserResponse<StreamChatGenerics>,
   ) {
-    const data = addFileToFormData(uri, name, contentType);
+    const data = addFileToFormData(uri, name, contentType || 'multipart/form-data');
     if (user != null) data.append('user', JSON.stringify(user));
 
-    return this.doAxiosRequest<SendFileAPIResponse>('post', url, data, {
+    return this.doAxiosRequest<SendFileAPIResponse>('postForm', url, data, {
       headers: data.getHeaders ? data.getHeaders() : {}, // node vs browser
       config: {
         timeout: 0,
@@ -1074,7 +1097,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   };
 
   /**
-   * Updates the members and watchers of the currently active channels that contain this user
+   * Updates the members, watchers and read references of the currently active channels that contain this user
    *
    * @param {UserResponse<StreamChatGenerics>} user
    */
@@ -1082,13 +1105,15 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     const refMap = this.state.userChannelReferences[user.id] || {};
     for (const channelID in refMap) {
       const channel = this.activeChannels[channelID];
-      /** search the members and watchers and update as needed... */
       if (channel?.state) {
         if (channel.state.members[user.id]) {
           channel.state.members[user.id].user = user;
         }
         if (channel.state.watchers[user.id]) {
           channel.state.watchers[user.id] = user;
+        }
+        if (channel.state.read[user.id]) {
+          channel.state.read[user.id].user = user;
         }
       }
     }
@@ -1460,7 +1485,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
    *
    * @param {BannedUsersFilters} filterConditions MongoDB style filter conditions
    * @param {BannedUsersSort} sort Sort options [{created_at: 1}].
-   * @param {BannedUsersPaginationOptions} options Option object, {limit: 10, offset:0}
+   * @param {BannedUsersPaginationOptions} options Option object, {limit: 10, offset:0, exclude_expired_bans: true}
    *
    * @return {Promise<BannedUsersResponse<StreamChatGenerics>>} Ban Query Response
    */
@@ -1590,8 +1615,8 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     query: string | MessageFilters<StreamChatGenerics>,
     options: SearchOptions<StreamChatGenerics> = {},
   ) {
-    if (options.offset && (options.sort || options.next)) {
-      throw Error(`Cannot specify offset with sort or next parameters`);
+    if (options.offset && options.next) {
+      throw Error(`Cannot specify offset with next`);
     }
     const payload: SearchPayload<StreamChatGenerics> = {
       filter_conditions: filterConditions,
@@ -1742,16 +1767,21 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
       throw Error(`Invalid channel group ${channelType}, can't contain the : character`);
     }
 
+    // support channel("messaging", {options})
+    if (channelIDOrCustom && typeof channelIDOrCustom === 'object') {
+      return this.getChannelByMembers(channelType, channelIDOrCustom);
+    }
+
+    // // support channel("messaging", undefined, {options})
+    if (!channelIDOrCustom && typeof custom === 'object' && custom.members?.length) {
+      return this.getChannelByMembers(channelType, custom);
+    }
+
     // support channel("messaging", null, {options})
     // support channel("messaging", undefined, {options})
     // support channel("messaging", "", {options})
-    if (channelIDOrCustom == null || channelIDOrCustom === '') {
+    if (!channelIDOrCustom) {
       return new Channel<StreamChatGenerics>(this, channelType, undefined, custom);
-    }
-
-    // support channel("messaging", {options})
-    if (typeof channelIDOrCustom === 'object') {
-      return this.getChannelByMembers(channelType, channelIDOrCustom);
     }
 
     return this.getChannelById(channelType, channelIDOrCustom, custom);
@@ -2253,23 +2283,21 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   /**
-   * _unblockMessage - unblocks message blocked by automod
+   * unblockMessage - unblocks message blocked by automod
    *
-   * Note: Do not use this.
-   * It is present for internal usage only.
-   * This function can, and will, break and/or be removed at any point in time.
    *
-   * @private
    * @param {string} targetMessageID
    * @param {string} [options.user_id] currentUserID, only used with serverside auth
    * @returns {Promise<APIResponse>}
    */
-  async _unblockMessage(targetMessageID: string, options: { user_id?: string } = {}) {
+  async unblockMessage(targetMessageID: string, options: { user_id?: string } = {}) {
     return await this.post<APIResponse>(this.baseURL + '/moderation/unblock_message', {
       target_message_id: targetMessageID,
       ...options,
     });
   }
+  // alias for backwards compatibility
+  _unblockMessage = this.unblockMessage;
 
   /**
    * @deprecated use markChannelsRead instead
@@ -2446,7 +2474,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   async updateMessage(
     message: UpdatedMessage<StreamChatGenerics>,
     userId?: string | { id: string },
-    options?: { skip_enrich_url?: boolean },
+    options?: UpdateMessageOptions,
   ) {
     if (!message.id) {
       throw Error('Please specify the message id when calling updateMessage');
@@ -2517,7 +2545,7 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
     id: string,
     partialMessageObject: PartialMessageUpdate<StreamChatGenerics>,
     userId?: string | { id: string },
-    options?: { skip_enrich_url?: boolean },
+    options?: UpdateMessageOptions,
   ) {
     if (!id) {
       throw Error('Please specify the message id when calling partialUpdateMessage');
@@ -2545,7 +2573,9 @@ export class StreamChat<StreamChatGenerics extends ExtendableGenerics = DefaultG
   }
 
   async getMessage(messageID: string) {
-    return await this.get<GetMessageAPIResponse<StreamChatGenerics>>(this.baseURL + `/messages/${messageID}`);
+    return await this.get<GetMessageAPIResponse<StreamChatGenerics>>(
+      this.baseURL + `/messages/${encodeURIComponent(messageID)}`,
+    );
   }
 
   getUserAgent() {